The 1d numerical density WCP prior

The 1d numerical density WCP prior

Usage

WCP_1D_Numerical_Density(
  W_func,
  eta,
  base_theta,
  L,
  U,
  mesh_width = NULL,
  mesh_n = 100,
  cutoff1 = 0.01,
  cutoff2 = NULL,
  L_included = FALSE,
  U_included = FALSE,
  inla_table = FALSE
)

Arguments

W_func

A function that returns the Wasserstein distance given theta.

eta

User specified parameter of the WCP prior.

base_theta

Base model theta.

L

Lower bound of theta.

U

Upper bound of theta.

mesh_width

Mesh width to be used. If not used, mesh_n will be used.

mesh_n

Number of mesh nodes to be used. Will not be used if mesh_width is given.

cutoff1

Cutoff parameter that decide upper bound of the Wasserstein distance on the right hand side of base_theta.

cutoff2

Cutoff parameter that decide upper bound of the Wasserstein distance on the left hand side of base_theta.

L_included

If L should be included in the domain, a TRUE or FALSE value

U_included

If U should be included in the domain, a TRUE or FALSE value

inla_table

Should the results be returned as a table to be used in INLA? Default is FALSE.

Value

A list of density locations and densities evaluated on those locations.