The 2d numerical approximation of WCP prior

The 2d numerical approximation of WCP prior

Usage

WCP_2D_Numerical_Density(
  W_func,
  eta,
  mesh_width,
  alpha = 1,
  tau = mesh_width/1000,
  cutoff = 0.01,
  region = list(type = "conic", lower_angle = 0, upper_angle = pi, base_theta = c(0, 0)),
  lc_multiplier = 20,
  buff_boundary = FALSE,
  parallel = FALSE,
  NumCores = parallel::detectCores() - 1,
  visual_mesh = FALSE
)

Arguments

W_func

A function that returns the Wasserstein distance given theta1 and theta2.

eta

User specified parameter of the WCP prior.

mesh_width

Mesh width.

alpha

Power of mesh width for a finer mesh.

tau

tau parameter, indicating how much bounded away from boundary.

cutoff

Cutoff parameter indicating probability tail mass for the Wasserstein distance.

region

A list that specify region information. For example, for strip regions (regions of the type IxJ, where I = (a,b) and J = (c,d) are intervals, with one of them being unbounded), one can set `region = list(type = "strip", base_theta = c(a,c), corners = c(a,b,c,d))`, where some of the values `a`,`b`,`c` or `d` can be `-Inf` or `Inf`. If the type is conic, the list must be set, for example, as list(type = 'conic', lower_angle = 0, upper_angle = pi, base_theta = c(0,0)), where `lower_angle` and `upper_angle` contains the smaller and larger angles that determine the cone.

lc_multiplier

Multiplier determines number of level curves

buff_boundary

Should an additional buffer (of the magnitude of mesh_width) be added to the region to bound it away from the boundary of the domain?

parallel

A logic value that indicating whether the user wants to run the function with multiple cpu.

NumCores

Number of cores to run the function.

visual_mesh

A logic value to determin whether to plot the mesh.

Value

A list of density locations, densities evaluated on those locations and other utilities that can help access accuracy.